+++ ssh_proxy_server/server.py
@@ -36,8 +36,8 @@ class SSHProxyServer:
 
         self.ssh_interface = ssh_interface
         self.scp_interface = scp_interface
-        self.sftp_interface = sftp_interface
         self.sftp_handler = sftp_handler
+        self.sftp_interface = self.sftp_handler.get_interface() or sftp_interface
         self.authentication_interface = authentication_interface
         self.authenticator = authenticator
         self.transparent = transparent
